+++ project/cheolhyun-onboarding/survey-domain/src/main/java/com/onboarding/model/response/IAnswerValue.java
@@ -1,4 +1,4 @@
-package com.onboarding.model;
+package com.onboarding.model.response;
 
 public interface IAnswerValue<T> {
     T getValue();
