+++ project/cheolhyun-onboarding/survey-domain/src/main/java/com/onboarding/model/response/MultipleChoiceAnswer.java
@@ -1,4 +1,4 @@
-package com.onboarding.model;
+package com.onboarding.model.response;
 
 import java.util.List;
